// packages/shared/src/sensor-types.ts
export interface GaugeThreshold {
  warn: number;
  crit: number;
  inverted?: boolean;
}

export type GaugeThresholds = Record<string, GaugeThreshold>;